chore: relocate typescript monorepo using pnpm
