+++ typescript/agentkit/.eslintrc.json
@@ -1,4 +1,4 @@
 {
   "parser": "@typescript-eslint/parser",
-  "extends": ["../../.eslintrc.base.json"]
+  "extends": ["../.eslintrc.base.json"]
 }
