+++ typescript/agentkit/package.json
@@ -40,6 +40,8 @@
   ],
   "dependencies": {
     "@coinbase/coinbase-sdk": "^0.17.0",
+    "abitype": "^1.0.8",
+    "decimal.js": "^10.5.0",
     "md5": "^2.3.0",
     "reflect-metadata": "^0.2.2",
     "twitter-api-v2": "^1.18.2",
