+++ typescript/examples/langchain-farcaster-chatbot/package.json
@@ -12,8 +12,8 @@
     "format:check": "prettier -c .prettierrc --check \"**/*.{ts,js,cjs,json,md}\""
   },
   "dependencies": {
-    "@coinbase/agentkit": "^0.1.0",
-    "@coinbase/agentkit-langchain": "^0.1.0",
+    "@coinbase/agentkit": "workspace:*",
+    "@coinbase/agentkit-langchain": "workspace:*",
     "@langchain/core": "0.3.30",
     "@langchain/langgraph": "^0.2.39",
     "@langchain/openai": "^0.3.16",
