+++ typescript/examples/langchain-farcaster-chatbot/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "preserveSymlinks": true,
     "outDir": "./dist",
